+++ demos/src/util/animating.rs
@@ -5,7 +5,7 @@ use bevy::utils::HashMap;
 #[derive(Component)]
 pub struct AnimationsHandler {
     pub player_entity: Entity,
-    pub animations: HashMap<String, Handle<AnimationClip>>,
+    pub animations: HashMap<String, AnimationNodeIndex>,
 }
 
 #[derive(Component)]
@@ -18,6 +18,7 @@ pub fn animation_patcher_system(
     parents_query: Query<&Parent>,
     scene_handlers_query: Query<&GltfSceneHandler>,
     gltf_assets: Res<Assets<Gltf>>,
+    mut animation_graphs_assets: ResMut<Assets<AnimationGraph>>,
     mut commands: Commands,
 ) {
     for player_entity in animation_players_query.iter() {
@@ -25,12 +26,24 @@ pub fn animation_patcher_system(
         loop {
             if let Ok(GltfSceneHandler { names_from }) = scene_handlers_query.get(entity) {
                 let gltf = gltf_assets.get(names_from).unwrap();
+                let mut graph = AnimationGraph::new();
+                let root_node = graph.root;
+                let mut animations = HashMap::<String, AnimationNodeIndex>::new();
+
+                for (name, clip) in gltf.named_animations.iter() {
+                    let node_index = graph.add_clip(clip.clone(), 1.0, root_node);
+                    animations.insert(name.to_string(), node_index);
+                }
+
                 let mut cmd = commands.entity(entity);
                 cmd.remove::<GltfSceneHandler>();
                 cmd.insert(AnimationsHandler {
                     player_entity,
-                    animations: gltf.named_animations.clone(),
+                    animations,
                 });
+                commands
+                    .entity(player_entity)
+                    .insert(animation_graphs_assets.add(graph));
                 break;
             }
             entity = if let Ok(parent) = parents_query.get(entity) {