+++ src/basis_action_traits.rs
@@ -20,8 +20,8 @@ pub struct TnuaBasisContext<'a> {
 
 impl TnuaBasisContext<'_> {
     /// The direction considered as "up".
-    pub fn up_direction(&self) -> Direction3d {
-        Direction3d::Y
+    pub fn up_direction(&self) -> Dir3 {
+        Dir3::Y
     }
 }
 
@@ -231,8 +231,8 @@ impl<'a> TnuaActionContext<'a> {
     }
 
     /// The direction considered as "up".
-    pub fn up_direction(&self) -> Direction3d {
-        Direction3d::Y
+    pub fn up_direction(&self) -> Dir3 {
+        Dir3::Y
     }
 }